+++ docs/examples/fourbar_linkage.py
@@ -1,6 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
+The fourbar_linkage module demonstrats the functionnalities of pylinkage.
+
+It is not intended to be imported in another project, but fell welcome to
+copy-paste chunks of code.
+
 Created on Sat Jun 19 12:32:37 2021.
 
 @author: HugoFara
@@ -39,51 +44,35 @@ pl.show_linkage(my_linkage)
 # We save initial position because we don't want a completely different movement
 init_pos = my_linkage.get_coords()
 
+
 @pl.kinematic_minimization
-def fitness_func(linkage, params, *args):
+def fitness_func(loci, **kwargs):
     """
     Return how fit the locus is to describe a quarter of circle.
 
     It is a minisation problem and the theorical best score is 0.
+
+    Returns
+    -------
+    float
+        Sum of square distances between tip locus bounding box and a defined
+        square.
     """
-    linkage.set_coords(init_pos)
-    linkage.set_num_constraints(params)
-    try:
-        points = 12
-        n = linkage.get_rotation_period()
-        # Complete revolution with 12 points
-        tuple(
-            tuple(i) for i in linkage.step(
-                iterations=points + 1, dt=n / points
-            )
-        )
-        # Again with n points, and at least 12 iterations
-        n = 96
-        factor = int(points / n) + 1
-        loci = tuple(
-            tuple(i) for i in linkage.step(
-                iterations=n * factor, dt=1 / factor
-            )
-        )
-    except pl.UnbuildableError:
-        return float('inf')
-    else:
-        # Locus of the Joint 'pin", mast in linkage order
-        tip_locus = tuple(x[-1] for x in loci)
-        # We get the bounding box
-        curr_bb = pl.bounding_box(tip_locus)
-        # Reference bounding box in order (min_y, max_x, max_y, min_x)
-        ref_bb = (0, 5, 2, 3)
-        # Our score is the square sum of the edges distances
-        return sum(
-            (pos - ref_pos) ** 2 for pos, ref_pos in zip(curr_bb, ref_bb)
-        )
+    # Locus of the Joint 'pin", mast in linkage order
+    tip_locus = tuple(x[-1] for x in loci)
+    # We get the bounding box
+    curr_bb = pl.bounding_box(tip_locus)
+    # Reference bounding box in order (min_y, max_x, max_y, min_x)
+    ref_bb = (0, 5, 2, 3)
+    # Our score is the square sum of the edges distances
+    return sum((pos - ref_pos) ** 2 for pos, ref_pos in zip(curr_bb, ref_bb))
+
 
 constraints = tuple(my_linkage.get_num_constraints())
 
 print(
     "Score before optimization: {}".format(
-        fitness_func(my_linkage, constraints)
+        fitness_func(my_linkage, constraints, init_pos)
     )
 )
 
@@ -110,7 +99,6 @@ bounds = pl.generate_bounds(my_linkage.get_num_constraints())
 
 score, position, coord = pl.particle_swarm_optimization(
     eval_func=fitness_func,
-    init_pos=init_pos,
     linkage=my_linkage,
     bounds=bounds,
     order_relation=min,